+++ tests/sanity/tests/actions.spec.ts
@@ -14,7 +14,7 @@ test.describe('actions tests', () => {
     await page.click('[id="app-recruit\\:string\\:RecruitApplication"]')
 
     await page.click('text=Talents')
-    await expect(page).toHaveURL(`${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/recruit%3Aapp%3ARecruit/talents`)
+    await expect(page).toHaveURL(`${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/recruit/talents`)
 
     await page.click('td:has-text("Frontend Engineer")')
 
@@ -29,30 +29,26 @@ test.describe('actions tests', () => {
     await page.click('[id="app-recruit\\:string\\:RecruitApplication"]')
 
     await page.click('text=Talents')
-    await expect(page).toHaveURL(`${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/recruit%3Aapp%3ARecruit/talents`)
+    await expect(page).toHaveURL(`${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/recruit/talents`)
 
     await page.press('body', 'Meta+k')
     await page.fill('[placeholder="type\\ to\\ filter\\.\\.\\."]', 'go to')
     expect(await page.locator('div.selectPopup :text("Go To Vacancies")').count()).toBe(1)
     await page.click('div.selectPopup :text("Go To Vacancies")')
 
-    await expect(page).toHaveURL(
-      `${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/recruit%3Aapp%3ARecruit/vacancies`
-    )
+    await expect(page).toHaveURL(`${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/recruit/vacancies`)
   })
   test('action-switch-applications', async ({ page }) => {
     await page.click('[id="app-recruit\\:string\\:RecruitApplication"]')
 
     await page.click('text=Talents')
-    await expect(page).toHaveURL(`${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/recruit%3Aapp%3ARecruit/talents`)
+    await expect(page).toHaveURL(`${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/recruit/talents`)
 
     await page.press('body', 'Meta+k')
     await page.fill('[placeholder="type\\ to\\ filter\\.\\.\\."]', 'go to')
     expect(await page.locator('div.selectPopup :text("Go To Applications")').count()).toBe(1)
     await page.click('div.selectPopup :text("Go To Applications")')
 
-    await expect(page).toHaveURL(
-      `${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/recruit%3Aapp%3ARecruit/candidates`
-    )
+    await expect(page).toHaveURL(`${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/recruit/candidates`)
   })
 })