+++ tests/sanity/tests/settings.spec.ts
@@ -17,14 +17,10 @@ test.describe('contact tests', () => {
     await page.click('.antiPopup-submenu >> text=Settings')
     // Click button:has-text("Setting")
     await page.click('button:has-text("Setting")')
-    await expect(page).toHaveURL(
-      `${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/setting%3Aids%3ASettingApp/setting`
-    )
+    await expect(page).toHaveURL(`${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/setting/setting`)
     // Click text=Edit profile
     await page.click('text=Edit profile')
-    await expect(page).toHaveURL(
-      `${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/setting%3Aids%3ASettingApp/profile`
-    )
+    await expect(page).toHaveURL(`${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/setting/profile`)
     // Click [placeholder="Location"]
     await page.click('[placeholder="Location"]')
     // Fill [placeholder="Location"]
@@ -50,11 +46,9 @@ test.describe('contact tests', () => {
     await page.click('.antiPopup-submenu >> text=Settings')
     // Click button:has-text("Templates")
     await page.click('button:has-text("Templates")')
-    await expect(page).toHaveURL(
-      `${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/setting%3Aids%3ASettingApp/message-templates`
-    )
-    // Go to http://localhost:8083/workbench%3Acomponent%3AWorkbenchApp/setting%3Aids%3ASettingApp/message-templates
-    await page.goto(`${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/setting%3Aids%3ASettingApp/message-templates`)
+    await expect(page).toHaveURL(`${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/setting/message-templates`)
+    // Go to http://localhost:8083/workbench%3Acomponent%3AWorkbenchApp/setting/message-templates
+    await page.goto(`${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/setting/message-templates`)
     // Click .flex-center.icon-button
     await page.click('#create-template >> .flex-center.icon-button')
     // Click [placeholder="New\ template"]
@@ -85,9 +79,7 @@ test.describe('contact tests', () => {
     await page.click('.antiPopup-submenu >> text=Settings')
     // Click button:has-text("Manage Statuses")
     await page.click('button:has-text("Manage Statuses")')
-    await expect(page).toHaveURL(
-      `${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/setting%3Aids%3ASettingApp/statuses`
-    )
+    await expect(page).toHaveURL(`${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/setting/statuses`)
     // Click text=Vacancies
     await page.click('text=Vacancies')
     // Click #create-template div