+++ tests/sanity/tests/workbench.spec.ts
@@ -12,27 +12,23 @@ test.describe('workbench tests', () => {
   test('navigator', async ({ page }) => {
     // Click [id="app-recruit\:string\:RecruitApplication"]
     await page.click('[id="app-recruit\\:string\\:RecruitApplication"]')
-    await expect(page).toHaveURL(`${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/recruit%3Aapp%3ARecruit`)
+    await expect(page).toHaveURL(`${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/recruit`)
     // Click text=Applications
     await page.click('text=Applications')
-    await expect(page).toHaveURL(
-      `${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/recruit%3Aapp%3ARecruit/candidates`
-    )
+    await expect(page).toHaveURL(`${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/recruit/candidates`)
     // Click text=Applications Application >> span
     await expect(page.locator('text=Applications Filter')).toBeVisible()
     await expect(page.locator('text="APP-1')).toBeDefined()
 
     // Click text=Talents
     await page.click('text=Talents')
-    await expect(page).toHaveURL(`${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/recruit%3Aapp%3ARecruit/talents`)
+    await expect(page).toHaveURL(`${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/recruit/talents`)
 
     await expect(page.locator('text=Andrey P.')).toBeVisible()
 
     // Click text=Vacancies
     await page.click('text=Vacancies')
-    await expect(page).toHaveURL(
-      `${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/recruit%3Aapp%3ARecruit/vacancies`
-    )
+    await expect(page).toHaveURL(`${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/recruit/vacancies`)
     // Click text=Software Engineer
     await page.click('text=Software Engineer')
     await expect(page.locator('text=Software Engineer')).toBeVisible()
@@ -42,7 +38,7 @@ test.describe('workbench tests', () => {
 
     // Click [id="app-chunter\:string\:ApplicationLabelChunter"]
     await page.click('[id="app-chunter\\:string\\:ApplicationLabelChunter"]')
-    await expect(page).toHaveURL(`${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/chunter%3Aapp%3AChunter`)
+    await expect(page).toHaveURL(`${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/chunter`)
 
     await page.click('text=general')
 
@@ -50,7 +46,7 @@ test.describe('workbench tests', () => {
     await expect(page.locator('.textInput')).toBeVisible()
 
     await page.click('[id="app-contact\\:string\\:Contacts"]')
-    await expect(page).toHaveURL(`${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/contact%3Aapp%3AContacts`)
+    await expect(page).toHaveURL(`${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/contact`)
     // Click text=John Appleseed
     await expect(page.locator('text=John Appleseed')).toBeVisible()
   })
@@ -62,7 +58,7 @@ test.describe('workbench tests', () => {
     await page.click('.antiPopup-submenu >> text=Settings')
     // Click button:has-text("Terms")
     await page.click('button:has-text("Terms")')
-    await expect(page).toHaveURL(`${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/setting%3Aids%3ASettingApp/terms`)
+    await expect(page).toHaveURL(`${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/setting/terms`)
     // Click .ac-header
     await expect(page.locator('.ac-header >> text=Terms')).toBeVisible()
   })